change cert etcd-client to etcd-server set etcd to TLS mode remove cloudinit bucket and cleanup it's policy refs

+++ resources/master/artifacts/upload/setup.sh
@@ -1,27 +1,32 @@
 #!/bin/bash -e
 
+GET=/opt/bin/s3get.sh
 # Get vault CA and PKI tokens
 $GET ${CONFIG_BUCKET} pki/ca.pem /opt/etc/vault/ca/ca.pem
-$GET ${CONFIG_BUCKET} pki-tokens/ca.pem /opt/etc/pki-tokens/etcd-client
 $GET ${CONFIG_BUCKET} pki-tokens/kube-apiserver /opt/etc/pki-tokens/kube-apiserver
+$GET ${CONFIG_BUCKET} pki-tokens/etcd-server /opt/etc/pki-tokens/etcd-server
+$GET ${CONFIG_BUCKET} etcd-init/initial-cluster /etc/sysconfig/initial-cluster
 
 $GET ${CONFIG_BUCKET} master/envvars /opt/etc/master/envvars
 source /opt/etc/master/envvars
 
 # Install vault binary
-docker run --rm -v /opt/bin:/tmp ${VAULT_IMAGE} cp /bin/vault /tmp/vault
-
+if [ ! -f /opt/etc/master/${VAULT_IMAGE} ]
+then
+  docker run --rm -v /opt/bin:/tmp ${VAULT_IMAGE} cp /bin/vault /tmp/vault
+  touch  /opt/etc/master/${VAULT_IMAGE}
+fi
 # Install kubernetes
-docker run --env VERSION="${KUBE_VERSION}" --rm -v /opt/bin:/shared xueshanf/install-kubernetes
+if [ ! -f /opt/etc/master/kube-${KUBE_VERSION} ]
+then
+  docker run --env VERSION="${KUBE_VERSION}" --rm -v /opt/bin:/shared xueshanf/install-kubernetes
+  touch /opt/etc/master/kube-${KUBE_VERSION}
+fi
 
 # Generate certs for etcd and kube-apiserver
 mkdir -p /var/lib/kubernetes/
-bash get-kube-certs.sh kube-apiserver etcd-client
+bash get-kube-certs.sh kube-apiserver etcd-server
 
 # Copy kube policy.jsonl and token.csv to /var/lib/kubernetes/
 # Note: token.csv will be copied to upload by "make upload-config"
 cp policy.jsonl token.csv /var/lib/kubernetes/
-
-# TODO: after install cp ${VAULT_RELEASE} and ${KUBE_VERSION} to local disk.
-# On rebooting compare ${VAULT_RELEASE} and ${KUBE_VERSION} with local copy
-# if matches do NOT install again
